backend/requirements.txt
django>=4.2
djangorestframework
django-cors-headers
dj-rest-auth
django-allauth
psycopg2-binary
python-dotenv
gunicorn
whitenoise
django-anymail[resend]
